Index working; templates working better
Partial templates for rows in list view are working and looking
good. Researched a little into possibly putting attachments or
the teams robot photo into couch itself as an attachment
so everything is nicely organized. Also getting ready to start
work on the big screen tomorrow: the scoreboard. Have to figure
out how to make it real time with limited time to write it...