start work on the scoreboard half, using reveal.js
Some fancy js and reveal.js to make a scrolling
sort of thingy for the scoreboard. Divs will update
with info when they are hidden ie: not the div being shown
at that moment, and will do so through a JSON request to
the API and then format and place the JSON into the div
for when it comes around to be shown. Thinking about doing
the same for changing which div is showing at the moment
so it could be changed through the web interface.